requirements.txt
PySide6>=6.8.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
pytz>=2024.1
